[node] Use the DefaultFileSource
We can remove a lot of code and make the bindings a lot simpler
by reusing the DefaultFileSource that is backed by curl. This
file source is widely tested by the [core] unit tests and used
by all the SDKs. The curl backend is used by the Linux port and
is known to be very stable.

This will also simplify the request logic. Previously the flow
for making a resource request was complicated to be implemented in
the bindings and had issues regarding object lifecycle.